# openapi/avedra-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AVEDRA API",
    "description": "API REST de AVEDRA — Asistencia, Convivencia y Evaluacion",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/usuarios": {
      "get": {
        "tags": [
          "usuarios"
        ],
        "summary": "Listar Usuarios",
        "description": "Lista usuarios con filtros opcionales (rol, activo).",
        "operationId": "listar_usuarios_api_v1_usuarios_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "rol",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rol"
            }
          },
          {
            "name": "activo",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Activo"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 25,
              "title": "Per Page"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_UsuarioResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "usuarios"
        ],
        "summary": "Crear Usuario",
        "description": "Crea un usuario nuevo.",
        "operationId": "crear_usuario_api_v1_usuarios_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NuevoUsuarioRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/usuarios/{usuario_id}": {
      "get": {
        "tags": [
          "usuarios"
        ],
        "summary": "Obtener Usuario",
        "description": "Obtiene un usuario por id.",
        "operationId": "obtener_usuario_api_v1_usuarios__usuario_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "usuario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Usuario Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "usuarios"
        ],
        "summary": "Actualizar Usuario",
        "description": "Actualiza nombre, email o telefono de un usuario.",
        "operationId": "actualizar_usuario_api_v1_usuarios__usuario_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "usuario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Usuario Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActualizarUsuarioRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/usuarios/{usuario_id}/desactivar": {
      "post": {
        "tags": [
          "usuarios"
        ],
        "summary": "Desactivar Usuario",
        "description": "Desactiva un usuario (soft delete).",
        "operationId": "desactivar_usuario_api_v1_usuarios__usuario_id__desactivar_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "usuario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Usuario Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/usuarios/{usuario_id}/reactivar": {
      "post": {
        "tags": [
          "usuarios"
        ],
        "summary": "Reactivar Usuario",
        "description": "Reactiva un usuario desactivado.",
        "operationId": "reactivar_usuario_api_v1_usuarios__usuario_id__reactivar_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "usuario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Usuario Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/usuarios/{usuario_id}/cambiar-rol": {
      "post": {
        "tags": [
          "usuarios"
        ],
        "summary": "Cambiar Rol",
        "description": "Cambia el rol de un usuario.",
        "operationId": "cambiar_rol_api_v1_usuarios__usuario_id__cambiar_rol_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "usuario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Usuario Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CambiarRolRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/usuarios/{usuario_id}/resetear-password": {
      "post": {
        "tags": [
          "usuarios"
        ],
        "summary": "Resetear Password",
        "description": "Resetea la contrasena de un usuario. Retorna la temporal si no se dio una explicita.",
        "operationId": "resetear_password_api_v1_usuarios__usuario_id__resetear_password_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "usuario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Usuario Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetearPasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActualizarUsuarioRequest": {
        "properties": {
          "nombre_completo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nombre Completo"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "telefono": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telefono"
          }
        },
        "type": "object",
        "title": "ActualizarUsuarioRequest"
      },
...
      "CambiarRolRequest": {
        "properties": {
          "nuevo_rol": {
            "type": "string",
            "title": "Nuevo Rol"
          }
        },
        "type": "object",
        "required": [
          "nuevo_rol"
        ],
        "title": "CambiarRolRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NuevoUsuarioRequest": {
        "properties": {
          "usuario": {
            "type": "string",
            "title": "Usuario"
          },
          "nombre_completo": {
            "type": "string",
            "title": "Nombre Completo"
          },
          "rol": {
            "type": "string",
            "title": "Rol"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "institucion_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Institucion Id"
          },
          "password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "usuario",
          "nombre_completo",
          "rol"
        ],
        "title": "NuevoUsuarioRequest"
      },
...
      "PaginatedResponse_UsuarioResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/UsuarioResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "per_page": {
            "type": "integer",
            "title": "Per Page"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "per_page"
        ],
        "title": "PaginatedResponse[UsuarioResponse]"
      },
...
      "ResetearPasswordRequest": {
        "properties": {
          "nueva_password": {
            "type": "string",
            "title": "Nueva Password",
            "default": ""
          }
        },
        "type": "object",
        "title": "ResetearPasswordRequest"
      },
...
      "UsuarioResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "usuario": {
            "type": "string",
            "title": "Usuario"
          },
          "nombre_completo": {
            "type": "string",
            "title": "Nombre Completo"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "telefono": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telefono"
          },
          "rol": {
            "type": "string",
            "title": "Rol"
          },
          "activo": {
            "type": "boolean",
            "title": "Activo"
          },
          "debe_cambiar_password": {
            "type": "boolean",
            "title": "Debe Cambiar Password"
          },
          "fecha_creacion": {
            "type": "string",
            "format": "date",
            "title": "Fecha Creacion"
          },
          "ultima_sesion": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ultima Sesion"
          },
          "institucion_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Institucion Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "usuario",
          "nombre_completo",
          "rol",
          "activo",
          "debe_cambiar_password",
          "fecha_creacion"
        ],
        "title": "UsuarioResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}